refactor(problem6): 크루 저장, 중복 이름을 가진 이메일을 찾는 비즈니스 로직을 별도의 서비스 레이어로 분리
